(Re)Introduce Embarcadero C++Builder XE3 compiler support.
win32.h: Define LOG4CPLUS_HAVE_C99_VARIADIC_MACROS for C++Builder XE3.

fileappender.cxx (get_locale_by_name): Avoid function try-catch.

lockfile.cxx: Use SH_DENYNO instead of _SH_DENYNO, etc., for C++Builder builds.

patternlayout.cxx: Use qualified call to formatRelativeTimestamp().

socket-win32.cxx: Silence signed/unsigned comparison warning.

+++ src/socket-win32.cxx
@@ -223,7 +223,7 @@ connectSocket(const tstring& hostn, unsigned short port, bool udp, SocketState& 
         INT ret = WSAStringToAddress (const_cast<LPTSTR>(hostn.c_str ()),
             AF_INET, 0, reinterpret_cast<struct sockaddr *>(&insock),
             &insock_size);
-        if (ret == SOCKET_ERROR || insock_size != sizeof (insock)) 
+        if (ret == SOCKET_ERROR || insock_size != static_cast<INT>(sizeof (insock))) 
         {
             state = bad_address;
             goto error;
